Remove draft chunk
